Issue #5338: moved, renamed and removed abstact declaration from AbstractLoader. Also moved and renamed AbstractLoaderTest and fixed suppressions for intellij idea and checkstyle
